Revert "vsock_proxy: Refactor DNS-related functionality"
This reverts commit 91b24f49e963aa9bbfa5c3a9e67fdc1442381b9b.
