+++ vsock_proxy/src/main.rs
@@ -10,7 +10,7 @@ use clap::{App, AppSettings, Arg};
 use env_logger::init;
 use log::info;
 
-use vsock_proxy::{starter::Proxy, IpAddrType, VsockProxyResult};
+use vsock_proxy::{starter::{Proxy, VsockProxyResult}, IpAddrType};
 
 fn main() -> VsockProxyResult<()> {
     init();
